Add template and input filtering support
This commit adds third-party support for rendering output via a template
layer, as well as for filtering input. This avoids the unnecessary
creation of a lot of custom code.

The libraries chosen are based upon the fact that I know them so well.
There may well be plenty of other libraries, but I'm very experienced
and comfortable with these ones.