Add a command to export the config sample file to AsciiDoc format
This new Command class implementation was originally based on the
original implementation, which converts the config sample file to
reStructuredText format. However, the more that I read the file, the
more I wanted to simplify the logic in it, and remove any unnecessary
code. That's why this second implementation is similar, yet distinctly
different.